Scripts/Python/requirements.txt
# Requirements for PBG Scripts
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0

# Asana SDK
asana>=5.2.0

# CLI UI
click>=8.0.0
rich>=14.0.0
textual>=0.1.0
trilium-py>=1.3.4

# Markdown processing
markdown>=3.0.0

# YAML processing
pyyaml>=6.0.0

# Date handling
python-dateutil>=2.8.0
